페이징, 상세페이지 들어갔다가 돌아오면은 들어가 페이지

+++ project_2025/src/main/java/com/project/page/board1/dao/ReplyDao.java
@@ -10,6 +10,6 @@ public interface ReplyDao {
 	
 	void insertReply(Reply reply);
 
-	long getNextGroupId();
+	int getNextGroupId();
 
 }
