+++ project_2025/src/main/java/com/project/page/board1/dao/ReplyDaoOracle.java
@@ -25,7 +25,7 @@ public class ReplyDaoOracle implements ReplyDao {
 	}
 
 	@Override
-	public long getNextGroupId() {
+	public int getNextGroupId() {
 		return sql.selectOne("board1.getNextGroupId");
 	}
